[backend] refactor: move users auth to app layer
